Move client provider to client pkg
